Bug fixed. NamedType with unsigned base: ZERO_VAL == NULL_VAL!

+++ EvolutionCoreInterface/EvolutionTypes.h
@@ -11,12 +11,12 @@
 #include "GeneType.h"
 #include "NamedType.h"
 
-using ENERGY_UNITS   = NamedType< short,          struct ENERGY_UNITS_Parameter >;
-using EVO_GENERATION = NamedType< long,           struct EVO_GENERATION_Parameter >;
-using IND_ID         = NamedType< long,           struct IND_ID_Parameter >;
-using MEM_INDEX      = NamedType< unsigned short, struct MEM_INDEX_Parameter >;
-using ACTION_COUNT   = NamedType< unsigned int,   struct ACTION_COUNT_Parameter >;
-using PERCENT        = NamedType< short,          struct PERCENT_Parameter >;
+using ENERGY_UNITS   = NamedType< short, struct ENERGY_UNITS_Parameter >;
+using EVO_GENERATION = NamedType< long,  struct EVO_GENERATION_Parameter >;
+using IND_ID         = NamedType< long,  struct IND_ID_Parameter >;
+using MEM_INDEX      = NamedType< short, struct MEM_INDEX_Parameter >;
+using ACTION_COUNT   = NamedType< long,  struct ACTION_COUNT_Parameter >;
+using PERCENT        = NamedType< short, struct PERCENT_Parameter >;
 
 constexpr PERCENT operator"" _PERCENT( unsigned long long ull )
 {
